requirements.txt
gspread==6.1.4
google-auth==2.34.0
